Improve game engine; refactor Camera to Player and implement collision structures
Renamed and modified the Camera class to create a Player class to make it more intuitive. Also, fundamental concepts for a future collision system have been implemented, including a count of solid voxels for chunks and the creation of Axis Aligned Bounding Boxes. Improved rendering efficiency by changing the storage of chunks from an array to a dictionary and considering the player's distance from a chunk while rendering. Added SqrMagnitude to the Vector3 struct.