Update math and rendering functionality, add async testing and main thread synchronization. Refactored terrain generation
This commit updates several components of the math and rendering functionality in the VoxelGame. It modifies Plane.cs to update vector and plane calculations, and it revises shadow mapping in ShadowMapper.cs. This commit also adds a new asynchronous testing function and a MainThreadSynchronization mechanism to improve thread safety. Moreover, shader files are updated and code to display wireframes is added. New files for rendering optimisations, GizmoBox and counter functionalities are included as well.